pyproject.toml
[project]
name = "jwt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.3",
    "fastapi[standard]>=0.123.5",
    "pyjwt>=2.10.1",
]
